grainConfig: take array of policies
__Context__
After implementing the `Recent` policy and a variant of `Balanced` (#2584), we can see that adding policy parameters for new policies is both tedious on from our perspective as well as for the user.

More specifically, on the user side, their config is a flat object with parameters from different policies sitting side by side.  Here, we implement more structure by taking an array of policies.

Right now, an example `config/grain.json` might look as follows:
```js
{
  "balancedPerWeek": 30
  "immediatePerWeek": 50,
  "recentPerWeek": 20,
  "maxSimultaneousDistributions": 2,
  "discount": 0.5
}
```

This also forces us to check each policy explicitly downstream in `toDistributionPolicy`, which could be a lot cleaner if we could leverage the `AllocationPolicy` type we already have.

Config now uses an array of allocation policies, with the option
of using deprecated parameters for now.

This might look as follows:
```js
{
  balancedPerWeek: 10, // (deprecated)
  recentPerWeek: 30, // (deprecated)
  recentWeeklyDecayRate: 0.5, // (deprecated)
  allocationPolicies: [
    {
      policyType: "BALANCED",
      budget: "100"
    },
    {
      policyType: "RECENT",
      budget: "50",
      discount: 0.1
    },
    {
      policyType: "RECENT",
      budget: "100",
      discount: 0.5
    }
}
```

This also lets us take multiple versions of the same policies.  This allows us to create richer ensembles of policies.  For instance, I might want to run a `Recent` policy with a 90% discount, and another one with a 5% discount.

__Implementation Details__
- Adds support for `opaque type NonnegativeGrain`, which is now used
by each policy as the type of their budget, reducing the need to handle
a few error cases.
- Each policy now has its own config parser, which is used to handle
human-friendly (i.e. 20g instead of 20000000...g).  This eliminates the
need to create intermediary config types and functions to handle transformations
between them.  Here, we can parse directly into `AllocationPolicy`'s.

__Test Plan__
Unit tests are provided here for both the parser, and well as `toDistributionPolicy`,
which serves to check the config.