MPG: Support retrieving the intervals (#2588)
This commit adds an `intervals` accessor to the MarkovProcessGraph which
allows retrieving the intervals. In order to get the end time of the
last interval, we add an extra numeric field to the JSON serialized form
(and the in memory representation).

I also added a new explicit failure case when there are no intervals.

Test plan: Unit tests added, `yarn test` passes.