core/weights: Weights class
`Weights` composes `NodeWeights` and `EdgeWeights` into a higher order
and more useful factory that also lets us interface with `NodeWeights`
and `EdgeWeights` directly, instead of needing to summon `Weights` and
only using half of it.

```js
// Weights: WeightsI
export type WeightsI = {
  fold: () => WeightsT,
  nodeWeightsT: () => NodeWeightsT,
  edgeWeightsT: () => EdgeWeightsT,
  getNodeWeight: (NodeAddressT) => NodeWeight,
  getEdgeWeight: (EdgeAddressT) => EdgeWeight,
  setNodeWeight: (NodeAddressT, NodeWeight) => WeightsI,
  setEdgeWeight: (EdgeAddressT, EdgeWeight) => WeightsI,
  merge: (
    ws: $ReadOnlyArray<WeightsI>,
    resolvers?: WeightsTResolvers
  ) => WeightsI,
  copy: () => WeightsI,
  toJSON: () => WeightsJSON,
};
```

This commit is also an effort to shape a cleaner experience handling
node and edge weights, particularly when using a `NodeWeightEvaluator`
or `EdgeWeightEvaluator`.

Currently, we cannot get the node or edge weight directly as it requires
transforming a Map into a Trie (which is performed by the `Node/EdgeWeightEvaluator`.
This class creates an interface for getting node and edge weights directly
from the class, letting it lazily handle contructing the trie.

```js
// Before
// weights: WeightsT
import { nodeWeightEvaluator } from "../***/weightEvaluator"
const nwe = nodeWeightEvaluator(weights);
const nodeWeight = nwe(address);

// After
// weights: Weights (class)
const nodeWeight = weights.nodeWeight(address);
```

__Design Considerations__
- `WeightsT` does the testing for the lower level `NodeWeightsT` and `EdgeWeightsT`.  While it makes sense to implement tests for each module, the tests themselves cover the logic in one aggregated swoop, so I decided to leave as is.
- Now, the `WeightsI`/`WeightsT` types can be a little confusing.  This commit is an in-place refactor, but once `WeightsI` replaces `WeightsT` in the remainder of the codebase, `WeightsT` will be hidden and/or removed entirely.

__Test Plan__
Unit tests provided for the class methods that test along the same conditions
that the `WeightsT` functions are tested.  For the `node/edgeWeight` methods,
tests are a superset of `weightEvaluator` tests, also checking for the lazy
creation of a `node/edgeWeightEvaluator` on the class.