Testing: Remove Snapshot from createGraph tests
This file fails the sharness test for file names in the describe string
when grep is aliased without the `-a` option. This option forces grep to
attempt to parse files, even if they are recognized as binary type.

The reason for this seems to be that in POSIX systems, any file
containing a null literal is recognized by default as a binary file.

Rather than utilizing a snapshot, the test now manually transforms the
edge properties into console and source-safe parts for comparison.
Additionally, a length check is added to ensure that likedPosts are
indeed tested and empty arrays aren't compared.

test plan: yarn unit