Update Discourse snapshots (#2624)
Discourse stats are a bit out of date in the snapshots.
Additionally it appears that some discourse endpoints have changed.

It's interesting that the full sharness snapshot tests fail locally for me regularly, but do not fail consistently in our PRs. I'm guessing this might be some kind of caching issue in CI, but I figure if updating the snapshots fixes things reliably locally, it's probably not worth the time investigating exactly what's going on in CI right now. Open to an alternative perspective though

test plan: yarn sharness-full