Add tags to the Discourse data model (#2626)
This commit updates the Discourse fetcher and mirror so that we now
retrieve the tags associated with a topic, and make them retrievable in
the mirror.

This change will not change Cred.

Test plan:
New unit tests confirm that tags are retrievable from the mirror
repository. The snapshot updates should reveal that tags are now fetched
from Discourse.

Part of #2455.