Spec out WeightedGraph2 external API (#2608)
WeightedGraph2 is the successor to WeightedGraph. It is
a class rather than a type, and as such, it abstracts away the
requirement of assigning every single node and edge (entity) each their own
weight values. With this class, arbitrary sets of nodes and edges can
all be assigned a weight with one assignment, by assigning an Address
prefix a weight.

An Address prefix is a subset of address parts that always contains
the address' root part as the first entry. The prefix may also include
an entity's entire address in this case, if a plugin wishes to update
an entity's weight after calling `addNode` or `addWeight`.

Entities may also have their weights assigned directly. If an
entity is passed in without a weight, it is assigned the default weight
of 1 (or 1 in both directions for edges).

This class will also be extensible to include tags in the future.

test plan: none; No logic is implemented. This is solely for the purpose
of API design review.