Update Save Ledger logic to use LedgerManager
Replaced all the instances of the manually written POST requests with a saveLedger function
so that the LedgerManager logic is used to save to disk and abstracted from the UI layer. Can
be used in the future to persist the ledger to various backends.

Also added an alert so the user gets a visual confirmation that the Ledger has been
saved (or if there was an error)

Test Plan: Update the ledger from the Identities / Transfer / Special distribution pages
and save it. Ensure the ledger file is updated and the alert confirms the ledger was saved.