Bugfix: Explicitly stringify Discourse CategoryIds
When refactoring Discourse CategoryIds to be strings instead of ints, I
was initially relying on the sqlite mirror to implicitly convert the
integers to strings. Testing on the sqlite console had returned expected
results. However, I had not considered the middleware we use to
interact with sqlite DB's, and in interation testing it as found that
the categoryId strings all contained an appended `.0`.

These appended zero-value decimals cause lookup mismatches downstream
when like-weights are calculated, and rather than applying the expected
weight, the default weight is applied.

For example, if I have the following configuration:
```json
categoryWeights: {
  "16": 5
}
```

the topic's categoryId loaded into memory from sqlite would be `16.0`,
and in `_categoryWeight` the lookup `weights.get(`16.0`)` will return
`undefined` and the defaultWeight will be assigned. This means the that
`weights.get(`16`) will never execute and the configured weight of 5
will never be applied.

By calling `toString` on the `category_id` api value as soon as we
receive the response from the discourse API, we bypass this issue, and
moreover, conform to flow's expectation that topic.categoryId is a
string immediately.

test plan: yarn unit

I considered writing some defensive tests for mirror and/or
mirrorRepository to capture the issue detailed above, but if we just
stringify the result immediately, I'm not sure it's worth writing them.