Bugfix: Explicitly stringify Discourse CategoryIds (#2657)
When refactoring Discourse CategoryIds to be strings instead of ints, I
was initially relying on the sqlite mirror to implicitly convert the
integers to strings. Testing on the sqlite console had returned expected
results. However, I had not considered the middleware we use to
interact with sqlite DB's, and in integration testing it was found that
the categoryId strings all contained an appended `.0`.

These appended zero-value decimals cause lookup mismatches downstream
when like-weights are calculated, and rather than applying the expected
weight, the default weight is applied.

For example, if I have the following configuration:
```json
categoryWeights: {
  "16": 5
}
```

the topic's categoryId loaded into memory from sqlite would be `16.0`,
and in `_categoryWeight` the lookup `weights.get(`16.0`)` will return
`undefined` and the defaultWeight will be assigned. This means the that
`weights.get(`16`) will never execute and the configured weight of 5
will never be applied.

By calling `toString` on the `category_id` api-returned value as soon as we
receive the response from the discourse API, we bypass this issue, and
moreover, conform to flow's expectation that topic.categoryId is a
string immediately.

# Test Plan
```bash
yarn unit
```
On a local instance of the sc's cred repo:
1. Optionally remove all plugins from sourcecred.json except for 
`sourcecred/discourse` to speed up testing.
2. Run `sourcecred.js go && sourcecred.js serve`and observe the 
current users' cred values.
3. Add category configs in the discourse config so it appears like so:
```json
{
  "serverUrl": "https://discourse.sourcecred.io",
  "weights": {
    "categoryWeights": {
      "13": 5,
      "26": 5
    }
  }
}
```
4. run `sourcecred.js go --no-load && sourcecred.js serve`
5. Observe the cred values increase

I considered writing some defensive tests for mirror and/or
mirrorRepository to capture the issue detailed above, but if we just
stringify the result immediately, I'm not sure it's worth writing them.

ht to @s-ben for finding this bug