cli/grain2: grain2 command using CredGraph
__Context__
This commit implements a `grain2` command, which is a simple fork of `grain` that
moves away from using the soon to be deprecated `CredView` and instead uses `CredGraph`.
This is part of a larger effort from @blueridger to replace `CredView` with `CredGrainView`.

Note that `grain2` now loads the `CredGraph` instead of `CredResults`.

This commit also implements forks of `applyDistributions` and `computeCredAccounts` to
take a `CredGraph` instead of a `CredView`.  We can phase out the originals when we
deprecate `CredView` officially.

__Assumptions__
- Original applyDistributions and computeCredAccounts works.  I forked this assuming
  it works, but there's no test for it.

__Test Plan__
Ran `grain2` and `grain` on the same sample instance.
- Distribution output in ledger matches using JSON.stringify
- -s flag does not write to ledger in grain2 as expected