core/ledger: add NonnegativeGrain type (#2666)
__Context__
This commit adds a `NonnegativeGrain` type which will be used
by grain allocation policies for the type of their budgets.  This helps
push a subspace of error handling onto Flow.

`NonnegativeGrain` is implemented as an oqaque type similar to Grain,
with stricter lifting functions which check for nonnegativity.

__Test Plan__
Unit tests for `NonnegativeGrain` are provided.  The tests are forks of
tests on `Grain`.