policies: Use NonnegativeGrain as budget type (#2669)
* core/ledger: add NonnegativeGrain type

__Context__
This commit adds a `NonnegativeGrain` type which will be used
by grain allocation policies for the type of their budgets.  This helps
push a subspace of error handling onto Flow.

`NonnegativeGrain` is implemented as an oqaque type similar to Grain,
with stricter lifting functions which check for nonnegativity.

__Test Plan__
Unit tests for `NonnegativeGrain` are provided.  The tests are forks of
tests on `Grain`.

* policies: Use NonnegativeGrain as budget type

__Context__
This commit uses the `NonnegativeGrain` type for the budget
type of grain allocation policies.  This helps strongly enforce the nonnegativity
of budgets, while reducing the need to manually check if budgets are negative.
In addition, we use `NonnegativeGrain`'s parser to parse directly into policies without any upgrading.

__Test Plan__
Policy unit tests all updated to lift budget input into `NonnegativeGrain`.