grain allocation policies: add config parser (#2670)
__Context__
This commit adds a new parser for each grain allocation policy,
as well as a higher order parser which checks that the data parses
to one of the allocation policies.

This will be used in our upgraded grain config, which will take an
array of allocation policies.  This parser parses user input directly
into allocation policies without needing any upgrade functions.  This
will help us eliminate the `toDistributionPolicy` function in
`api/grainConfig` entirely.

The main difference in this parser is that it allows users to enter
human readable grain amounts instead of attoGrain amounts without
needing to write an upgrade function.

__Test Plan__
Trivial change, tests for the `numberParser` subparser are provided
in `core/ledger/nonnegativegrain.test`.