Implement LedgerDiskStorage for LedgerManager (#2653)
This implementation of LedgerStorage can be used in a locally running frontend to replace
our current method for loading and saving the ledger to the local filesystem. This storage
adapter can later be swapped out for one that talks to GitHub directly, allowing for instance
management without dealing with git and eventually without needing to run it locally once we
have our auth system in place

Test Plan: Dont have unit tests in frontend code, so just sanity check that the implementation
looks right. Future PR will refactor the loading code to use this instead, that's when it can
be tested in the front end directly.