grainConfig: take array of policies (#2671)
__Context__
After implementing the Recent policy and a variant of Balanced (#2584), we can see that adding policy parameters for new policies is both tedious on from our perspective as well as for the user.

More specifically, on the user side, their config is a flat object with parameters from different policies sitting side by side. Here, we implement more structure by taking an array of policies.

Right now, an example config/grain.json might look as follows:

```js
{
  "balancedPerWeek": 30
  "immediatePerWeek": 50,
  "recentPerWeek": 20,
  "maxSimultaneousDistributions": 2,
  "discount": 0.5
}
```

This also forced us to check each policy explicitly downstream in toDistributionPolicy, which could be a lot cleaner if we could leverage the AllocationPolicy type we already have. Now, we have a parser which will eliminate the need for the
function entirely after we remove the deprecated config.

Config now looks (for example) as follows:

```js
{
  allocationPolicies: [
    {
      policyType: "BALANCED",
      budget: 100
    },
    {
      policyType: "RECENT",
      budget: 50,
      discount: 0.1
    },
    {
      policyType: "RECENT",
      budget: 100,
      discount: 0.5
    }]
}
```

Note that for the time being, we support the deprecated config.  For example:
```js
{
  balancedPerWeek: 10,
  immediatePerWeek: 10,
  allocationPolicies: [
    {
      policyType: "BALANCED",
      budget: 100
    },
    {
      policyType: "RECENT",
      budget: 50,
      discount: 0.1
    },
    {
      policyType: "RECENT",
      budget: 100,
      discount: 0.5
    }]
}
``

This also lets us take multiple versions of the same policies. This allows us to create richer ensembles of policies. For instance, I might want to run a Recent policy with a 90% discount, and another one with a 5% discount.

__Test Plan__
Unit tests are provided here for both the parser, and well as toDistributionPolicy, which serves to check the config.