Core: Implement WeightedGraph2
WeightedGraph2 is a class that encapsulates WeightsI and Graph. It
provides an API that is similar to Graph in both form and expected
behavior, except it returns a calculated weight along with each node
and edge.

It also allows address prefixes to have weights assigned, which allows
for weight values to compound beyond an addresses' configured weight.
This behavior is actually provided via the algorithm's WeightsEvaluator,
but this API allows for more intuitive configuration

test plan: yarn unit

Tests cover the entire class and verify the functionality described here
and in the docstrings