Make numIntervalsLookback field in Immediate policy config optional
Ledgers that had Immediate policy distributions before the numIntervalsLookback feature was added
don't have this field in their logs, which results in a runtime error that breaks backwards compatibility.
This fixes it by making this field optional and defaulting it to 1 when it's not defined.

Test Plan: Ensure CI tests pass and ledgers with old Immediate distributions load without errors