Setup separate API builds for web and node environments
The current API build is designed to only work in web browsers and not a NodeJS environment in the backend.
In order to allow API access to SourceCred from any environment, this fix creates two separate webpack
builds for web and node environments so that things like `fetch` are correctly polyfilled for the intended
environment. Previously trying to import and use the SourceCred API in a node environment would cause a
fatal error.

Test Plan: Ensure `yarn build:api` runs correctly and two separate builds are created in dist/server/api.js
and dist/client/api.js.