Fix createIdentity in discord plugin and add tests
Previously if the `member.nick` field was undefined or an empty string, the createIdentity function
would throw a "Cannot read properly 'slice' of undefined" error instead of falling back to the username.

Additionally, the code was previously untested so I added tests to ensure correctness in the future.

Test Plan: Ensure CI tests pass