Bug fix: observe discord rate-limits (#2738)
There was an issue when loading data from discord API, where we would make endless requests exceeding our rate-limit. This caused inconsistent cred scores several times a day on big instances. The proposed solution adds a new attribute to the fetcher called timeout. After a valid response, we check the discord response headers and see if we have any requests remaining before hitting the rate-limit. If not, the timeout variable is set to the time when we are allowed to make more requests.

Every time the fetcher is called, it checks if rate limit timeout has been set by Discord's API and waits for the timeout to pass before making any further requests. This design would allow for parallel use of the class, since all fetch requests must observe a the timeout.

test plan:
In the tests, we covered the scenario where the fetcher would get a rate-limited response and wait until it can continue.
We also ran integration tests with the SourceCred Production instance.

Co-authored-by: Kevin Siegler <17910833+topocount@users.noreply.github.com>