Implement PayableAddressStore for Ledger Accounts (#2429)
PayableAddressStore maps addresses of cryptocurrencies to a participant's
address capable of accepting the currency. This structure exists to
accommodate safe migration for grain/payout token changes. 
In a future front-end, users will need to verify themselves that the address
they are supplying is capable of receiving their share of a grain distribution.

An address may be deleted by passing in an empty string for the
`address` parameter. This is useful in case the underlying private key
is compromised or the exchange hosting a custodial account is hacked.

The Currency objects are currently serialized before use in the Map.
This is due to the fact that Objects used as keys must be directly
accessible in order to retrieve the mapped value. A new object with
matching members is not equal to the object used as the key because each
has a distinct memory reference. By stringifying the object before
entering it as the key, deep equality can be observed and the
composition of the object becomes the keys and values it contains, 
rather than the object's memory reference.

test plan: Unit tests are provided