update docs for ledger.lastDistributionTimestamp (#2745)
The API for ledger.lastDistributionTimestamp doesn't specify what
happens when there haven't been any distributions. Counter-intuitively,
we return -Infinity, and we actually depend on that behavior elsewhere,
so it can't easily be changed. This commit just updates the
documentation so that we track this accordingly, and also I filed a new
issue to consider changing this.

Test plan: `yarn test` passes, since it's only an update to docs/test
code, this definitely suffices.