Temp fix for chart bug so a line is displayed for single interval
Currently, the interval for the top chart is set to a week. As such,
when viewing "last week" or "this week" the line only drew one point
and therefore didn't display a line. This is a temporary fix so
single weeks are viewable, and so the "current month" and "all time"
lines fill the entire chart.