Redirect target IdentityIds to map to base Account
Because there is little chance of collision as future identities are
created, it is safe to redirect target IdentityIds to point to the base
account in the `Ledger.prototype._accounts` Map.

A unit test has been added to ensure that old IdentityIds don't lose
their connection to latest baseAccount.

It'd be preferable to track identity merge history at some point, but
for now, this lightweight implementation will allow for sucessfull
account lookups when rendering older grain distributions and other events
that preceded a merge.

fixes #2753

test plan:
- Unit tests updated
- Follow the replication steps in #2753 to ensure no error is thrown in
react admin or the browser console