Implement and utilize NetworkStorage
This class is a DataStorage wrapper for basic fetch GET calls.
It throws with the HTTP error when the response is not "OK". This allows
for I/O helpers to handle different error codes accordingly.

test plan:
unit test are provided. the ui/load.js file has also been updated and
works with the snapshotted instance and a local SourceCred instance.