Implement DiskStorage using DataStorage Interface (#2776)
DiskStorage is a low-level class intended for I/O with a local disk. It
implements the WritableDataStorage interface, which is meant to allow
for uniform abstractions for any potential I/O the codebase interacts
with.

Payloads are specified as byte strings. This type is meant to enforce the
abstract representation of blobs. Uint8Arrays are chosen as the
underlying type because they are supported in both the browser and
NodeJS. Additionally, they cannot be confused for builtin JavaScript
strings.

test plan: unit tests are provided