Utilize ZipStorage
Description
-

For now, this class will be used to compress weighted graphs and cred
graphs. these artifacts aren't especially readable on disk and also take
up a 10s of MBs. This will allow larger instances to continue utilizing
Github for the foreseeable future. If further optimization is needed, a
higher-order ShardedStorage class can be implemented and used.

Test Plan
-
`yarn build`

in a local instance:
1. `scDev go`
2. `scDev grain`
3. `scDev serve`

open http://localhost:6006 and verify that the frontends load as
expected.