Bugfix: Reinstantiate CredValue for mutation
Description
-
In explorer timeline component, we attempt to mutate a prop.
Reinstantiation of the value within the component allows the mutation to
occur, and the render succeeds.

The issue occurs specifically when The credValue array is of length 1,
    and we attempt to add a second entry in the array, presumably to
    form a line.

Test Plan
-
Find some data that that has CredView array lengths(s) of 1. I used the
user-provided data here:
https://discord.com/channels/453243919774253079/718263631158050896/822135748014112798

paste the output and data folders into your instance of choice, and
serve it. Ensure explorer home loads.