Specify a Grain Integration Contract
This PR contains two artifacts that should be implemented by Grain
Integrations.

The first is the `IntegrationConfig`. The integration provides the
expected config values. These config values should then be checked
against the current ledger and instance config to ensure the integration
can execute without error.

The second is the `grainIntegration` function interface. This is
interface accepts a map of payout addresses and grain amounts, and a
Currency type, which contains either a protocol indicator or token
address if the protocol is the EVM-based. If the function returns an
`IntegrationResult` and accounting is enabled, the ledger is updated with
the new balances.

Functions that wrap the grain integration are implemented to show how it
is meant to be used, and are meant to be helpers once this feature is in
production.

Test Plan

Unit tests are provided for the helper functions.

The helper functions are meant to illustrate how grain integrations will be
used.