Set up grain distribution tracking for grain integrations (#3003)
# Semantics

A "tracked distribution" is a distribution that has been created when
distribution tracking is enabled

Tracked distributions are catalogued as "not executed" upon creation, and
only when `markDistributionExecuted` is called, will they be marked as "executed".
Note that `markDistributionExecuted` does not itself executed integration
functionality. This exists outside of the ledger context for now.

# Further Details

Grain integration tracking will allow us to track which distributions are
eligible to be processed by a grain integration. This will be needed for
situations where a grain integration is actually transacting funds, such as
on a blockchain. Grain integrations that generate some form of output that
does not directly relate to a transaction can still execute, but they
may be executed multiple times to generate the report in different
contexts, so there is no need to track these executions in the ledger.

# Test Plan
unit tests are provided.