Fix identity name reference detector
When referring to user names in initiatives, the existing reference detector converts the string
in the initiative JSON to lowercase but doesnt convert the names in the ledger to lowercase, which
causes the reference detector to not match with an identity when the name in the ledger has uppercase
letters.

This fixes it by converting all the names read from the ledger to lowercase before comparing with the
reference.

Test Plan: Ensure new unit tests pass