Implement CSV Grain Integration Plugin
This plugin outputs a csv file. Each row contains a payout address in
the first column and the raw amount to be paid in the second column.

grain integrations shall be prefixed in their package.json files with
`sc-grainIntegration-*` this will allow for grouping and filtering using
lerna and allow for a clean change in the scope of concerns.

The output is optimized for https://disperse.app

test plan: unit tests are provided.