Set up Grain Integration support in Config API (#3079)
This PR sets up a grain config for grain integrations. The config is validated against a `bundledGrainIntegrations` allow-list. and the mapped integration function is then set as a part of the config.

The ledger currency parser is now utilized to allow for token and some other crypto-related configs to be set in an instance and utilized by the ledger when setting payout addresses. However, this config is not directly utilized by the ledger yet

Finally, The grain integration interface has been updated to accept the entire config. This is done to allow for verification that the instance is configured correctly by the instance, and potentially error if a config is incompatible with the invoke grain integration.

test plan: unit tests have been updated