Utilize an Opaque type to key on currencies in the Ledger (#3231)
By utilizing an opaque type, we can eliminate the details of
stringifying the Currency objects from the ledger. This allows for some
cleaner scope of concern.

The opaque type prevents the decoding of keys, which
should not be necessary, Generating a key should be a one-way operation,
in case we ever move to generating hashes instead of just stringified
objects, or some other operation to that changes the characteristics of
keys.

Additionally, json-stable-stringify is utilized over JSON.stringify
because it produces the same string across all platforms. Some
JSON.stringify implementations differ and therefore it is potentially
unstable.

test plan: unit tests have been updated