Feature: Disable balance accounting in the ledger
When utilizing an on chain grain integration, tracking account balances
in the sourcecred ledger makes little sense. These changes disable
balance accounting, while preserving the functionality key to
SourceCred, which is tracking distributions across identities. While
balances are no longer tracked, paid totals will continue to be. This
feature, combined with Distribution execution tracking, narrows the
scope of the ledgers focus to the state of distributions, and whether or
not they have been executed outside of SourceCred.

When account balance tracking is disabled, the following changes are
made to ledger behavior:
- all grain balances are zeroed out in mutable accounts tracked by the
  ledger
- grain transfers are disabled
- distributed grain is not added to a user's balance. the paid total is
  updated, however. Undistributed balances are tracked as a whole through
  the distribution tracking functionality added earlier.
- identities without a relevant payout address will be deactivated
- identities that have no relevant payout addresses configured cannot be
  activated. relevant means the external currency configured in the
  ledger's state.

Test plan: Unit tests included