Migrate to a maintained MerkleTree library (#3186)
Using 'merkletreejs' will allow us to utilize the latest, most secure
MerkleTree JS implementations. This is a widely used library for many
crypto projects. The API is very similar, but the construction is much
more versatile and requires some specific configuration, as reflected in
the changes to tests.

test plan: unit tests have been updated to utilize this new dependency.