Author Pre-Audit Merkle Distribution Contract
This smart contract is very similar to Balancer's merkle-redeem smart
contract, plus the addition of a feature that allows for a delay before
allocations are live to be claimed.
The smart contract has a role system that allows for delegation of
allocation publishing and pausing roles.

This added functionality allows for a more decentralized workflows and
potentially automation, since the powers of the account publishing
allocations are tightly scoped, and there is recourse between the delay
period and pausability.

It might be worth adding a `minimum_delay` period that must be observed
by the `SEEDER_ROLE` holder and can be conrolled by `PAUSER_ROLE`
holders.

Test Plan:
Unit tests for the smart contract are included.
The smart contract is fully covered, though coverage instrumentation
hasn't been set up.