feat: [wip] adds support for importing custom user defined plugins
This is to help users define custom plugins, without needing to
use external plugins. It allows the developer to specify any package
to use as a plugin, as long as it implements the `Plugin` interface.

The motivation here is to lower the barrier to entry for extending
sourcecred.

This is still a wip so lacks a few tests.