combo bugfix: `spacer` option renamed to `space` (#3280)
`spacer` is not an option in `json-stable-stringify`
and therefore has no effect on the formatting of strings passed in. As
demonstrated in the updated unit tests, spacing is now applied to the
expected strings.

Additionally instead of finding a way to correctly render an empty
array, a more verbose, human readable error messsage is supplied when
the empty case throws. I also added a contextual prefix to improve the
readability of both the empty and non-empty case.

test plan:
Unit tests updated to demonstrate the correct indentation.